Script for receiving data from WordPress page v0.8
This script must be on the database server side.
- Requirements:
        * A Web server ( like Apache )
        * Be publicly accessible to an address known to the administrator of the WordPress site.
          ( for example : http://donationbox.database.org/index.php )

This script receives the data that the WordPress site sends, and (for now) writes them in the file "data.log".
In a future release (οnce the database has been finalized), instead of logging the data into a file,
it should store them in the main database ("donationbox_network").